split test window components into seperate files
