+++ packages/client/src/components/project/results/ExportMenu.tsx
@@ -6,7 +6,7 @@ import { GetTestjobs_testjobs } from "./__generated__/GetTestjobs";
 import ViewButton from "./ViewButton";
 import TestButton from "./TestButton";
 import RenameExportButton from "./RenameExportButton";
-import TestHistory from "./TestHistory";
+import TestWindow from "./TestWindow";
 
 export default function ExportMenu(props: {
   exprt: GetProjectData_project_exports;
@@ -78,7 +78,7 @@ export default function ExportMenu(props: {
         <RenameExportButton id={props.exprt.id} handler={handleClose} />
         {testButton}
         {viewButton}
-        <TestHistory exprt={props.exprt} handler={handleClose} />
+        <TestWindow exprt={props.exprt} handler={handleClose} />
       </Menu>
     </>
   );
